refactor(V2Resolver): extract the SQL resolver to a concern
