+++ app/controllers/v2/supported_profiles_controller.rb
@@ -11,7 +11,7 @@ module V2
     private
 
     def supported_profiles
-      @supported_profiles ||= authorize(resolve_collection)
+      @supported_profiles ||= authorize(fetch_collection)
     end
 
     def resource
